requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
qdrant-client>=1.7.0
openai>=1.10.0
spacy>=3.7.0
pypdf>=3.17.0
python-docx>=1.1.0
unstructured[email]>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0